+++ backend/src/main/java/com/ddbb/dingdong/domain/transportation/simulator/subscription/publisher/BusPublisherFactory.java
@@ -1,6 +1,6 @@
-package com.ddbb.dingdong.simulator.subscription.publisher;
+package com.ddbb.dingdong.domain.transportation.simulator.subscription.publisher;
 
-import com.ddbb.dingdong.simulator.subscription.BusSubscriptionManager;
+import com.ddbb.dingdong.domain.transportation.simulator.subscription.BusSubscriptionManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.geo.Point;
 import org.springframework.stereotype.Service;
