+++ backend/src/test/java/com/ddbb/dingdong/bus/BusSimulatorTest.java
@@ -1,13 +1,12 @@
 package com.ddbb.dingdong.bus;
 
-import com.ddbb.dingdong.simulator.subscription.BusSubscriptionManager;
-import com.ddbb.dingdong.simulator.subscription.UserSubscription;
-import com.ddbb.dingdong.simulator.subscription.publisher.BusPublisherFactory;
-import com.ddbb.dingdong.simulator.subscription.source.BusSimulator;
-import com.ddbb.dingdong.simulator.segment.RouteSegmentProvider;
-import com.ddbb.dingdong.simulator.segment.impl.StubRouteSegmentProvider;
-import com.ddbb.dingdong.simulator.BusSimulatorFactory;
-import com.ddbb.dingdong.simulator.subscription.subscriber.StubConsoleSubscriber;
+import com.ddbb.dingdong.domain.transportation.simulator.subscription.BusSubscriptionManager;
+import com.ddbb.dingdong.domain.transportation.simulator.subscription.UserSubscription;
+import com.ddbb.dingdong.domain.transportation.simulator.subscription.publisher.BusPublisherFactory;
+import com.ddbb.dingdong.domain.transportation.simulator.segment.RouteSegmentProvider;
+import com.ddbb.dingdong.domain.transportation.simulator.segment.impl.StubRouteSegmentProvider;
+import com.ddbb.dingdong.domain.transportation.simulator.BusSimulatorFactory;
+import com.ddbb.dingdong.domain.transportation.simulator.subscription.subscriber.StubConsoleSubscriber;
 import org.junit.jupiter.api.Test;
 import org.springframework.data.geo.Point;
